refactor: components로 파일 정리
