+++ apps/web/src/app/[locale]/guild/(components)/GuildCard.tsx
@@ -1,13 +1,16 @@
 'use client';
 
+import { useState } from 'react';
 import Image from 'next/image';
 import { Flex } from '_panda/jsx';
 import { flex } from '_panda/patterns';
 import { type Guild } from '@gitanimals/api';
+import { inboxQueries } from '@gitanimals/react-query';
 import { useQueryClient } from '@tanstack/react-query';
 import { UsersRoundIcon } from 'lucide-react';
 
 import { Link } from '@/i18n/routing';
+import { joinGuildAction } from '@/serverActions/guild';
 
 interface GuildCardProps {
   guild: Guild;
@@ -15,6 +18,20 @@ interface GuildCardProps {
 
 export function GuildCard({ guild }: GuildCardProps) {
   const queryClient = useQueryClient();
+  const [isJoinPetSelectOpen, setIsJoinPetSelectOpen] = useState(false);
+
+  const submitJoinGuild = async (selectPersona: string) => {
+    try {
+      await joinGuildAction({
+        guildId: guild.id,
+        personaId: selectPersona,
+      });
+
+      queryClient.invalidateQueries({ queryKey: inboxQueries.allKey() });
+    } catch (error) {
+      console.error(error);
+    }
+  };
 
   return (
     <>
