+++ apps/web/src/app/[locale]/guild/(components)/GuildDetail.tsx
