+++ apps/web/src/app/[locale]/guild/(components)/GuildSearch.tsx
